rename: 커스텀 예외를 implement.support.exception으로 이동
