+++ src/main/kotlin/com/yourssu/scouter/common/implement/domain/department/DepartmentReader.kt
@@ -1,5 +1,6 @@
 package com.yourssu.scouter.common.implement.domain.department
 
+import com.yourssu.scouter.common.implement.support.exception.DepartmentNotFoundException
 import org.springframework.stereotype.Component
 import org.springframework.transaction.annotation.Transactional
