+++ src/main/kotlin/com/yourssu/scouter/hrms/implement/support/exception/MemberNotFoundException.kt
@@ -1,4 +1,4 @@
-package com.yourssu.scouter.hrms.implement.domain.member
+package com.yourssu.scouter.hrms.implement.support.exception
 
 class MemberNotFoundException(
     override val message: String
